Removed color BLACK from properties
The "NZXT H9 Elite ATX Mid Tower White with Tempered Glass Side Panel and USB 3.2 Gen 2x2 Type-C, USB 3.2 Gen 1 Type-A" as its name says is "WHITE"! There is not a single cm2 of it that is black and it is annoying that it shows up first when filtering for "BLACK" pc cases.